fix(web): show the Toss maximum alongside other providers, and correct the ceiling comments
Two follow-ups from a final review of the quota ceiling change.

The Toss maximum notice was gated on usesTossAmountMode, which is only true when
Toss is the *sole* payment path. Enabling Toss together with Stripe, epay or any
other provider left the notice unrendered -- so the feature closed the gap only
for Toss-exclusive installs. It now renders whenever Toss is offered at all: the
ceiling is a fixed KRW figure, so stating it is correct in every configuration.

The over-limit *warning* stays narrower on purpose, still firing only in the
Toss-only amount mode with KRW selected. That is the one case where the typed
number is unambiguously the KRW charge. In quota mode the charge is derived
server-side from group ratio and discount, and when Toss shares the form with
other providers the input is not necessarily KRW at all -- warning there would
compare against a number we do not have.

Separately, comments introduced by the ceiling change claimed it was enforced by
"request validation for admin grants". That is false: ManageUser add_quota and
override validate only the sign. Worse, widening users.quota to bigint removed the
out-of-range rejection MySQL and PostgreSQL used to supply on that path, which was
its only de facto bound. SQLite was always unbounded there, so the deployment is
unaffected, and it takes root to reach -- a footgun, not an exploit. The comments
in common/constants.go and model/user.go and the doc now describe the gap instead
of asserting coverage; closing it is deferred and tracked in the doc.

Also drops a now-redundant int64() cast on an already-int64 field, and updates two
leftover references to the abandoned 9e15 ceiling (a tier-table comment and a test
boundary case, now 5e14 = common.MaxQuota).

+++ model/wallet_auto_recharge.go
@@ -3667,7 +3667,7 @@ func walletThresholdShouldCharge(tx *gorm.DB, policy *WalletAutoRecharge, now ti
 		if err := tx.Select("quota").Where("id = ?", policy.TargetId).First(&user).Error; err != nil {
 			return false, err
 		}
-		return int64(user.Quota) <= policy.ThresholdQuota, nil
+		return user.Quota <= policy.ThresholdQuota, nil
 	}
 }
